get rid of some Borland compiler warnings
[git-p4: depot-paths = "//depot/qt/main/": change = 74199]

+++ src/tools/qsettings.cpp
@@ -1835,7 +1835,7 @@ QStringList QSettings::readListEntry(const QString &key, bool *ok )
 	    esc = TRUE;
 	} else {
 	    s.append(value[i]);
-	    if ( i == value.length()-1 )
+	    if ( i == (int)value.length()-1 )
 		l.append(s);
 	}
     }
